Improved page covers
The performance of the blur was quite awful, and just removing it made
the gradient ugly.
Using a mask to fade the image does the same job while being much less
intensive.
`postcss-easing-gradients` would have been great to smooth out the
mask's gradient but isn't flexible enough for this use case, as it
doesn't support custom stops.
I would have put together a separate module, but not tonight.